Formated with forge fmt

+++ test/FlashLoanExample.t.sol
@@ -7,14 +7,13 @@ import "aave-v3-core/contracts/interfaces/IPoolAddressesProvider.sol";
 import "@openzeppelin/contracts/interfaces/IERC20.sol";
 
 contract FlashLoanExampleTest is Test {
-   //Ethereum Mainnet DAI Contract Address
+    //Ethereum Mainnet DAI Contract Address
     address constant DAI = 0x6B175474E89094C44Da98b954EedeAC495271d0F;
 
     // Mainnet Pool contract address
-    address constant POOL_ADDRESS_PROVIDER =
-        0x2f39d218133AFaB8F2B819B1066c7E434Ad94E9e;
+    address constant POOL_ADDRESS_PROVIDER = 0x2f39d218133AFaB8F2B819B1066c7E434Ad94E9e;
 
-        //vm.envString is provided by Foundry so that we can read our .env file and get values from there    
+    //vm.envString is provided by Foundry so that we can read our .env file and get values from there
     string MAINNET_RPC_URL = vm.envString("MAINNET_RPC_URL");
 
     uint256 mainnetFork;
@@ -23,21 +22,19 @@ contract FlashLoanExampleTest is Test {
     FlashLoanExample public flashLoanExample;
 
     function setUp() public {
-                // vm is a variable included in the forge standard library that is used to manipulate the execution environment of our tests
-        // create a fork of Ethereum mainnet using the specified RPC URL and store its id in mainnetFork 
+        // vm is a variable included in the forge standard library that is used to manipulate the execution environment of our tests
+        // create a fork of Ethereum mainnet using the specified RPC URL and store its id in mainnetFork
         mainnetFork = vm.createFork(MAINNET_RPC_URL);
         // select the fork obtained using its id
         vm.selectFork(mainnetFork);
-         //deploy FlashLoanExample to the created fork with POOL_ADDRESS_PROVIDER as its constructor argument
-        flashLoanExample = new 
-    FlashLoanExample(IPoolAddressesProvider(POOL_ADDRESS_PROVIDER));
-         //fetch the DAI contract
+        //deploy FlashLoanExample to the created fork with POOL_ADDRESS_PROVIDER as its constructor argument
+        flashLoanExample = new FlashLoanExample(IPoolAddressesProvider(POOL_ADDRESS_PROVIDER));
+        //fetch the DAI contract
         token = IERC20(DAI);
-
     }
 
     function testTakeAndReturnLoan() public {
-         // Get 2000 DAI in our contract by using deal
+        // Get 2000 DAI in our contract by using deal
         // deal is a cheatcode that lets us arbitrarily set the balance of any address and works with most ERC-20 tokens
         uint256 BALANCE_AMOUNT_DAI = 2000 ether;
         deal(DAI, address(flashLoanExample), BALANCE_AMOUNT_DAI);
@@ -45,10 +42,9 @@ contract FlashLoanExampleTest is Test {
         // Request and execute a flash loan of 10,000 DAI from Aave
         flashLoanExample.createFlashLoan(DAI, 10000);
 
-                // By this point, we should have executed the flash loan and paid back (10,000 + premium) DAI to Aave
+        // By this point, we should have executed the flash loan and paid back (10,000 + premium) DAI to Aave
         // Let's check our contract's remaining DAI balance to see how much it has left
-        uint remainingBalance = token.balanceOf(address(flashLoanExample));
-
+        uint256 remainingBalance = token.balanceOf(address(flashLoanExample));
 
         // Our remaining balance should be <2000 DAI we originally had, because we had to pay the premium
         //asserLt => assert strictly less than
